+++ packages/browser-core/src/transport/httpRequest.spec.ts
@@ -1,5 +1,5 @@
-import type { EndpointBuilder, HttpRequestEvent  } from '@datadog/js-core/transport'
-import { createEndpointBuilder } from '@datadog/js-core/transport'
+import type { EndpointBuilder, HttpRequestEvent } from '@datadog/js-core/transport'
+import { createEndpointBuilder, RECOMMENDED_REQUEST_BYTES_LIMIT } from '@datadog/js-core/transport'
 import type { Request } from '../../test'
 import {
   collectAsyncCalls,
@@ -12,7 +12,7 @@ import {
 } from '../../test'
 import { noop } from '../tools/utils/functionUtils'
 import type { HttpRequest } from './httpRequest'
-import { createHttpRequest, fetchStrategy, RECOMMENDED_REQUEST_BYTES_LIMIT } from './httpRequest'
+import { createHttpRequest, fetchStrategy } from './httpRequest'
 
 describe('httpRequest', () => {
   const ENDPOINT_URL = 'http://my.website'
